backend/requirements.txt
# =========================
# FLASK STACK
# =========================
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.1

# =========================
# DATABASE & STORAGE
# =========================
pymongo>=4.6.0
# motor>=3.3.0            # Uncomment if you plan to use Asynchronous MongoDB

# =========================
# AUTHENTICATION
# =========================
bcrypt>=4.1.0
PyJWT>=2.8.0

# =========================
# DATA PROCESSING
# =========================
pandas>=2.2.0
numpy>=1.26.0

# =========================
# MACHINE LEARNING
# =========================
scikit-learn>=1.4.0
joblib>=1.3.0

# =========================
# UTILITIES
# =========================
python-multipart>=0.0.9   # Required for file uploads

gunicorn>=21.2.0